refactor Server methods into several sourcce files; select() implemented in server loop
