+++ include/Server.hpp
@@ -5,10 +5,16 @@
 # include <map>
 # include <sys/select.h> // for fd_set
 
-class	User;	//no include needed as only pointer is used
+class	User;	// no include needed as only pointer is used
 
 class Server
 {
+	public:
+		Server(int port, const std::string& password);
+		~Server();
+
+		void	run();
+
 	private:
 		// Disable default constructor and copying (makes no sense for a server)
 		Server();
@@ -16,26 +22,23 @@ class Server
 		Server&	operator=(const Server& other);
 
 		int								_port;		// Server port
-		int								_fd;		// Listening socket fd
+		int								_fd;		// server socket fd (listening socket)
 		std::string						_password;	// Server password
 		std::map<int, User*>			_usersFd;	// Keep track of active users by fd
 		std::map<std::string, User*>	_usersNick;	// Keep track of active users by nickname
 
-		void		initSocket();
-		int			prepareReadSet(fd_set& readFds);
-		void		acceptNewUser();
+		// === ServerSocket.cpp ===
 
-	public:
-		Server(int port, const std::string& password);
-		~Server();
+		void	initSocket();
+		int		prepareReadSet(fd_set& readFds);
 
-		void		run();	// Starts the server loop, accepting incoming connections and managing users
+		// === ServerUser.cpp ===
 
-		void		deleteUser(int fd);
-		void		deleteUser(const std::string& nickname);
-
-		User*		getUser(int fd) const;
-		User*		getUser(const std::string& nickname) const;
+		void	acceptNewUser();
+		void	deleteUser(int fd);
+		void	deleteUser(const std::string& nickname);
+		User*	getUser(int fd) const;
+		User*	getUser(const std::string& nickname) const;
 };
 
 # endif
